set in magic box is by now unsigned long long, mod magic box stub draw id which draw randoms from 1 - 50

+++ headers/magicBox.hpp
@@ -7,7 +7,7 @@
 
 class MagicBox {
 private:
-	std::set<unsigned> idBox_{};
+	std::set<unsigned long long> idBox_{};
 public:
 	std::size_t getIdBoxSize() const {
 		return idBox_.size();
